Attempt #2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int divider = toBeDivided / 5;
+        float divider = toBeDivided / 5;
         System.out.println(divider);
     }
 }
